new file structure, extracted "CNN" to one file
