+++ src/main_grad_cam.py
@@ -1,8 +1,6 @@
 # initial code from https://medium.com/@myringoleMLGOD/simple-convolutional-neural-network-cnn-for-dummies-in-pytorch-a-step-by-step-guide-6f4109f6df80
 
-import random
 import torch
-import torch.nn.functional as F
 from torch import optim
 from torch import nn
 from torch.utils.data import DataLoader
@@ -17,57 +15,7 @@ import cv2  # Import OpenCV for resizing
 from PIL import Image
 import time
 import os
-
-
-class CNN(nn.Module):
-    def __init__(self, in_channels, num_classes=10):
-        super(CNN, self).__init__()
-        # First convolutional layer: 1 input channel, 8 output channels, 3x3 kernel, stride 1, padding 1
-        self.conv1 = nn.Conv2d(
-            in_channels=in_channels, out_channels=8, kernel_size=3, stride=1, padding=1
-        )
-        # Max pooling layer: 2x2 window, stride 2
-        self.pool = nn.MaxPool2d(kernel_size=2, stride=2)
-        # Second convolutional layer: 8 input channels, 16 output channels, 3x3 kernel, stride 1, padding 1
-        self.conv2 = nn.Conv2d(
-            in_channels=8, out_channels=16, kernel_size=3, stride=1, padding=1
-        )
-        # Fully connected layer: 16*7*7 input features (after two 2x2 poolings), 10 output features (num_classes)
-        self.linearLayer1 = nn.Linear(16 * 7 * 7, num_classes)
-
-    def forward(self, x):
-        x = F.relu(self.conv1(x))
-        x = self.pool(x)
-        x = F.relu(self.conv2(x))
-        x = self.pool(x)
-        x = x.reshape(x.shape[0], -1)
-        x = self.linearLayer1(x)
-        return x
-
-
-def check_accuracy(loader, model, train):
-    if train:
-        print("Checking accuracy on training data")
-    else:
-        print("Checking accuracy on test data")
-
-    num_correct = 0
-    num_samples = 0
-    model.eval()
-
-    with torch.no_grad():
-        for x, y in loader:
-            x = x.to(device)
-            y = y.to(device)
-
-            scores = model(x)
-            _, predictions = scores.max(1)
-            num_correct += (predictions == y).sum()
-            num_samples += predictions.size(0)
-        accuracy = float(num_correct) / float(num_samples) * 100
-        print(f"Got {num_correct}/{num_samples} with accuracy {accuracy:.2f}%")
-    model.train()
-
+from model import CNN
 
 if __name__ == "__main__":
     device = "cuda" if torch.cuda.is_available() else "cpu"
@@ -93,7 +41,7 @@ if __name__ == "__main__":
     model = CNN(in_channels=1, num_classes=num_classes).to(device)
 
     if load_model == True:
-        model.load_state_dict(torch.load("saved_models/" + model_path))
+        model.load_state_dict(torch.load("saved_models/" + model_path, map_location=device))
         print(f"Model: {model} loaded from file")
     else:
         print(f"Model: {model}")
@@ -116,8 +64,8 @@ if __name__ == "__main__":
                 loss.backward()
                 optimizer.step()
 
-    check_accuracy(train_loader, model, train=True)
-    check_accuracy(test_loader, model, train=False)
+    model.check_accuracy(train_loader, train=True, device=device)
+    model.check_accuracy(test_loader, train=False, device=device)
 
     if load_model == False:
         torch.save(model.state_dict(), "saved_models/" + model_path)
