+++ src/model.py
@@ -1,0 +1,56 @@
+
+import torch
+import torch.nn.functional as F
+from torch import optim
+from torch import nn
+from torch.utils.data import DataLoader
+from tqdm import tqdm
+from load import ImageDataset
+from confusion_matrix import display_confustion_matrix_plot
+import numpy as np
+
+class CNN(nn.Module):
+    def __init__(self, in_channels, num_classes=10):
+        super(CNN, self).__init__()
+        # First convolutional layer: 1 input channel, 8 output channels, 3x3 kernel, stride 1, padding 1
+        self.conv1 = nn.Conv2d(in_channels=in_channels, out_channels=8, kernel_size=3, stride=1, padding=1)
+        self.relu1 = nn.ReLU()
+        # Max pooling layer: 2x2 window, stride 2
+        self.pool = nn.MaxPool2d(kernel_size=2, stride=2)
+        # Second convolutional layer: 8 input channels, 16 output channels, 3x3 kernel, stride 1, padding 1
+        self.conv2 = nn.Conv2d(in_channels=8, out_channels=16, kernel_size=3, stride=1, padding=1)
+        self.relu2 = nn.ReLU()
+        # Fully connected layer: 16*7*7 input features (after two 2x2 poolings), 10 output features (num_classes)
+        self.linearLayer1 = nn.Linear(16 * 7 * 7, num_classes)
+
+    def forward(self, x):
+        x = F.relu(self.conv1(x))
+        x = self.pool(x)
+        x = F.relu(self.conv2(x))
+        x = self.pool(x)
+        x = x.reshape(x.shape[0], -1)
+        x = self.linearLayer1(x)
+        return x
+
+    def check_accuracy(self, loader, train, device):
+        if train:
+            print("Checking accuracy on training data")
+        else:
+            print("Checking accuracy on test data")
+
+        num_correct = 0
+        num_samples = 0
+        self.eval()
+
+        with torch.no_grad():
+            for x, y in loader:
+                x = x.to(device)
+                y = y.to(device)
+
+                scores = self(x)
+                _, predictions = scores.max(1)
+                num_correct+=(predictions==y).sum()
+                num_samples+=predictions.size(0)
+            accuracy = float(num_correct) / float(num_samples) * 100
+            print(f"Got {num_correct}/{num_samples} with accuracy {accuracy:.2f}%")
+        self.train()
